CONTACT MANAGEMENT

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc create a contact
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log(`The request body is `, req.body);
@@ -28,6 +28,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(201).json(contact);
@@ -35,7 +36,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc get a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -48,13 +49,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update a contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permetions to update other contacts");
   }
 
   const updateContact = await Contact.findByIdAndUpdate(
@@ -68,12 +74,17 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete a contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permetions to delete other contacts");
   }
 
   await Contact.findByIdAndDelete(req.params.id);
